step 1
getpart.c needs to keep curl_memory.h
